register_function_cb(): take arg count for the instrumented function
For now, we use this to detect when we can't add a post-function hook
safely without corrupting the arguments passed to the instrumented
function via the stack. In the future, we may use this argument to
determine the size of data on the stack that we need to preserve to
allow instrumenting these functions.